add banglanews.club in for phishing
Facebook post for phishing attempt: https://archive.is/OXTWQ

The link https://lp.banglanews.club/ when loaded from browser it displays a login screen with some kind of login screen for betting a site but when archiving (https://archive.is/wBq9f) it loads a dummy news page? What are they trying to hide?